% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pwt_harmonized}
\alias{pwt_harmonized}
\title{A data set of harmonized PWT data}
\format{
A data frame of rgdpe data.
}
\usage{
pwt_harmonized
}
\description{
A data frame containing harmonized PWT data.
}
\examples{
pwt_harmonized
}
\keyword{datasets}
